+++ src/Vector/BLF/AfdxFrame.h
@@ -28,7 +28,6 @@
 
 #include <Vector/BLF/AbstractFile.h>
 #include <Vector/BLF/ObjectHeader.h>
-#include <Vector/BLF/VectorTypes.h>
 
 #include <Vector/BLF/vector_blf_export.h>
 
@@ -45,27 +44,27 @@ struct VECTOR_BLF_EXPORT AfdxFrame final : ObjectHeader {
 
     void read(AbstractFile & is) override;
     void write(AbstractFile & os) override;
-    DWORD calculateObjectSize() const override;
+    uint32_t calculateObjectSize() const override;
 
     /**
      * Ethernet (MAC) address of source computer
      * (network byte order).
      */
-    std::array<BYTE, 6> sourceAddress {};
+    std::array<uint8_t, 6> sourceAddress {};
 
     /**
      * The channel of the frame.
      */
-    WORD channel {};
+    uint16_t channel {};
 
     /**
      * Ethernet (MAC) address of target computer
      * (network byte order).
      */
-    std::array<BYTE, 6> destinationAddress {};
+    std::array<uint8_t, 6> destinationAddress {};
 
     /** enumeration for dir */
-    enum Dir : WORD {
+    enum Dir : uint16_t {
         /** Receive */
         Rx = 0,
 
@@ -79,7 +78,7 @@ struct VECTOR_BLF_EXPORT AfdxFrame final : ObjectHeader {
     /**
      * @brief Direction flag
      */
-    WORD dir {};
+    uint16_t dir {};
 
     /**
      * EtherType which indicates protocol for
@@ -88,28 +87,28 @@ struct VECTOR_BLF_EXPORT AfdxFrame final : ObjectHeader {
      * See Ethernet standard specification for valid
      * values.
      */
-    WORD type {};
+    uint16_t type {};
 
     /**
      * TPID when VLAN tag valid, zweo when no
      * VLAN. See Ethernet stnadard specification.
      */
-    WORD tpid {};
+    uint16_t tpid {};
 
     /**
      * TCI when VLAND tag valid, zero when no
      * VLAN. See Ethernet standard specification.
      */
-    WORD tci {};
+    uint16_t tci {};
 
     /**
      * Channel number of the underlying Ethernet
      * interface, where the frame originated from.
      */
-    BYTE ethChannel {};
+    uint8_t ethChannel {};
 
     /** reserved */
-    BYTE reservedAfdxFrame1 {};
+    uint8_t reservedAfdxFrame1 {};
 
     /**
      * Status- and error flags as:
@@ -125,16 +124,16 @@ struct VECTOR_BLF_EXPORT AfdxFrame final : ObjectHeader {
      * - Bit 8: A / B interface mismatch
      * - Bit 11: Fragmentation error
      */
-    WORD afdxFlags {};
+    uint16_t afdxFlags {};
 
     /** reserved */
-    WORD reservedAfdxFrame2 {};
+    uint16_t reservedAfdxFrame2 {};
 
     /**
      * Time period since last received frame on this
      * virtual link in micro-seconds
      */
-    ULONG bagUsec {};
+    uint32_t bagUsec {};
 
     /**
      * @brief Number of valid payLoad bytes
@@ -142,13 +141,13 @@ struct VECTOR_BLF_EXPORT AfdxFrame final : ObjectHeader {
      * Length of Ethernet payload data in bytes. Max.
      * 1500 Bytes (without Ethernet header)
      */
-    WORD payLoadLength {};
+    uint16_t payLoadLength {};
 
     /** reserved */
-    WORD reservedAfdxFrame3 {};
+    uint16_t reservedAfdxFrame3 {};
 
     /** reserved */
-    DWORD reservedAfdxFrame4 {};
+    uint32_t reservedAfdxFrame4 {};
 
     /**
      * @brief Ethernet payload data
